Extract functions and types from ChangeLevel capability
Related to #44
