+++ src/hubitat-capabilities/types/level-change-direction.type.ts
@@ -1,0 +1,9 @@
+/**
+ * @packageDocumentation
+ * @module HubitatCapabilities
+ */
+
+/**
+ * Possible directions of the level change.
+ */
+export type ELevelChangeDirection = 'down' | 'up';
